Stability of adding set methods
Attempted to use an inner class following the template method pattern
where field access getter and field access setter are two subclasses of
an abstract class (the template). But I think it is too complicated and
not worth the trouble for trying to minimize code. So copying the code
that generates field access getter for field access setter is easier to
code up, maintain, and stays close to the output of Bytecode Outline
Eclipse plugin.

- Add Commons Lang library for CamelCase utility
- Generate the name of ClassAccess methods by field type instead of
hardcoded strings
- Rename method parameter to obj from object
- Add ClassAccess interface methods for setting primitive types
- Add ClassAccess interface methods for setting primitive wrapper types