Make method access with parameters work
Was trying to solve a problem where a fatal error occurs and crashes the
JVM when a method was called with at least one parameter. The call(T
obj, int methodIndex) method was working i.e. the method access with
zero parameters. Turns out that I had forgotten to ALOAD the extra
parameters in the method access bridge.

- Create new AssignableInfo abstract class for have FieldInfo and
PropertyInfo extend from it
- Move internalName and type fields from MemberInfo into AssignableInfo
- Make MethodInfo extend MemberInfo. This is useful for to re-use
descriptor as method descriptor and name as method name. And to re-use
the table switch methods
- Test calling method with one parameter and another with no parameters
with void return type